Moved everything under util file for backward compatibility
