+++ streaming/base/format/base/reader.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass
 from typing import Any, Dict, Iterator, List, Optional, Set, Union
 
 from streaming.base.array import Array
-from streaming.base.utils import bytes_to_int
+from streaming.base.util import bytes_to_int
 
 __all__ = ['FileInfo', 'Reader', 'JointReader', 'SplitReader']
