+++ streaming/base/storage/upload.py
@@ -17,7 +17,7 @@ import tqdm
 
 from streaming.base.storage.download import (BOTOCORE_CLIENT_ERROR_CODES,
                                              GCS_ERROR_NO_AUTHENTICATION)
-from streaming.base.utils import get_import_exception_message
+from streaming.base.util import get_import_exception_message
 
 __all__ = [
     'CloudUploader',
