+++ streaming/base/util.py
@@ -3,10 +3,13 @@
 
 """Utility and helper functions for datasets."""
 
+import collections.abc
+import functools
 import os
+import random
 from multiprocessing.shared_memory import SharedMemory as BuiltinSharedMemory
 from time import sleep, time
-from typing import List, Union
+from typing import Any, Callable, List, Sequence, Type, TypeVar, Union, cast, overload
 
 import torch.distributed as dist
 
@@ -14,9 +17,11 @@ from streaming.base.constant import SHM_TO_CLEAN
 from streaming.base.distributed import get_local_rank, maybe_init_dist
 from streaming.base.shared.prefix import _get_path
 
+TCallable = TypeVar('TCallable', bound=Callable)
+
 __all__ = [
     'get_list_arg', 'wait_for_file_to_exist', 'bytes_to_int', 'number_abbrev_to_int',
-    'clean_stale_shared_memory', 'get_import_exception_message'
+    'clean_stale_shared_memory', 'get_import_exception_message', 'retry'
 ]
 
 
@@ -199,3 +204,93 @@ def get_import_exception_message(package_name: str, extra_deps: str) -> str:
     return f'Streaming was installed without {package_name} support. ' + \
             f'To use {package_name} related packages with Streaming, run ' + \
             f'`pip install \'mosaicml-streaming[{package_name}]\'`.'
+
+
+@overload
+def retry(
+    exc_class: Union[Type[Exception], Sequence[Type[Exception]]] = ...,
+    num_attempts: int = ...,
+    initial_backoff: float = ...,
+    max_jitter: float = ...,
+) -> Callable[[TCallable], TCallable]:
+    ...
+
+
+@overload
+def retry(exc_class: TCallable) -> TCallable:
+    # Use the decorator without parenthesis
+    ...
+
+
+# error: Type "(TCallable@retry) -> TCallable@retry" cannot be assigned to type "(func: Never) -> Never"
+def retry(  # type: ignore
+    exc_class: Union[TCallable, Type[Exception], Sequence[Type[Exception]]] = Exception,
+    num_attempts: int = 3,
+    initial_backoff: float = 1.0,
+    max_jitter: float = 0.5,
+):
+    """Decorator to retry a function with backoff and jitter.
+
+    Attempts are spaced out with
+    ``initial_backoff + 2**num_attempts + random.random() * max_jitter`` seconds.
+
+    Example:
+    .. testcode::
+
+        from streaming.base.util import retry
+
+        num_tries = 0
+
+        @retry(RuntimeError, num_attempts=3, initial_backoff=0.1)
+        def flaky_function():
+            global num_tries
+            if num_tries < 2:
+                num_tries += 1
+                raise RuntimeError("Called too soon!")
+            return "Third time's a charm."
+
+        print(flaky_function())
+
+    .. testoutput::
+
+        Third time's a charm.
+
+    Args:
+        exc_class (Type[Exception] | Sequence[Type[Exception]]], optional): The exception class or
+            classes to retry. Defaults to Exception.
+        num_attempts (int, optional): The total number of attempts to make. Defaults to 3.
+        initial_backoff (float, optional): The initial backoff, in seconds. Defaults to 1.0.
+        max_jitter (float, optional): The maximum amount of random jitter to add. Defaults to 0.5.
+
+            Increasing the ``max_jitter`` can help prevent overloading a resource when multiple
+            processes in parallel are calling the same underlying function.
+    """
+    if num_attempts < 1:
+        raise ValueError('num_attempts must be at-least 1')
+
+    def wrapped_func(func: TCallable) -> TCallable:
+
+        @functools.wraps(func)
+        def new_func(*args: Any, **kwargs: Any):
+            i = 0
+            while True:
+                try:
+                    return func(*args, **kwargs)
+                except exc_class as e:
+                    if i + 1 == num_attempts:
+                        raise e
+                    else:
+                        sleep(initial_backoff * 2**i + random.random() * max_jitter)
+                        i += 1
+
+        return cast(TCallable, new_func)
+
+    if not isinstance(exc_class, collections.abc.Sequence) and not (isinstance(
+            exc_class, type) and issubclass(exc_class, Exception)):
+        # Using the decorator without (), like @retry_with_backoff
+        func = cast(TCallable, exc_class)
+        exc_class = Exception
+
+        return wrapped_func(func)
+
+    return wrapped_func
