--- streaming/base/utils/__init__.py
@@ -1,19 +1,0 @@
-# Copyright 2023 MosaicML Streaming authors
-# SPDX-License-Identifier: Apache-2.0
-
-"""Helper utilities."""
-
-from streaming.base.utils.retrying import retry
-from streaming.base.utils.util import (bytes_to_int, clean_stale_shared_memory,
-                                       get_import_exception_message, get_list_arg,
-                                       number_abbrev_to_int, wait_for_file_to_exist)
-
-__all__ = [
-    'get_list_arg',
-    'wait_for_file_to_exist',
-    'bytes_to_int',
-    'number_abbrev_to_int',
-    'clean_stale_shared_memory',
-    'get_import_exception_message',
-    'retry',
-]
