+++ streaming/text/convert/enwiki_text.py
@@ -10,7 +10,7 @@ from typing import List
 from tqdm import tqdm
 
 from streaming.base import MDSWriter
-from streaming.base.utils import get_list_arg
+from streaming.base.util import get_list_arg
 
 
 def parse_args() -> Namespace:
