+++ streaming/vision/convert/ade20k.py
@@ -12,7 +12,7 @@ from typing import Any, Dict, Iterable, List, Tuple
 from tqdm import tqdm
 
 from streaming.base import MDSWriter
-from streaming.base.utils import get_list_arg
+from streaming.base.util import get_list_arg
 
 
 def parse_args() -> Namespace:
