+++ streaming/vision/convert/coco.py
@@ -15,7 +15,7 @@ from torch.utils.data import Dataset
 from tqdm import tqdm
 
 from streaming.base import MDSWriter
-from streaming.base.utils import get_list_arg
+from streaming.base.util import get_list_arg
 
 
 def parse_args() -> Namespace:
